Exit Buttons and Early Display Development
I added an easy way for the user to quickly exit the program from any open window.

I added a display for the connect 4 board to the SinglePlayer and TwoPlayer windows. This was done using System.Drawing and I had to extensive research to get it to work.

Let me know what you guys think.